+++ src/widgets/widget.ts
@@ -4,7 +4,7 @@ import GLib from 'gi://GLib?version=2.0';
 import Gdk from 'gi://Gdk?version=3.0';
 import Cairo from 'gi://cairo?version=1.0';
 import Service, { Props, BindableProps, Binding } from '../service.js';
-import { registerGObject, kebabify } from '../utils/gobject.js';
+import { registerGObject, kebabify, type CtorProps } from '../utils/gobject.js';
 import { interval, idle } from '../utils.js';
 import { Variable } from '../variable.js';
 import { App } from '../app.js';
@@ -79,7 +79,7 @@ type Bind = [
     transform?: (value: any) => any,
 ];
 
-export interface CommonProps<Attr> {
+interface CommonProps<Attr> {
     class_name?: string
     class_names?: Array<string>
     click_through?: boolean
@@ -92,9 +92,10 @@ export interface CommonProps<Attr> {
 
 export type BaseProps<Self, Props, Attr = unknown> = {
     setup?: (self: Self) => void
-} & BindableProps<Props & CommonProps<Attr>>
+} & BindableProps<CtorProps<Props & CommonProps<Attr>>>
 
-export interface Widget<Attr> extends CommonProps<Attr> {
+type Requierd<T> = { [K in keyof T]-?: T[K] };
+export interface Widget<Attr> extends Requierd<CommonProps<Attr>> {
     hook<
         Gobject extends GObject.Object,
     >(