+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 - add: Window.gdkmonitor
 - export modules globally
 - make Audio.microphone and Audio.speaker always
+- feat: greetd service (#282)
+- feat(pam): Utils.authenticate (#273)
 - feat: child property as second parameter [#265](https://github.com/Aylur/ags/pull/265/)
 
 ## Breaking change